resolving a problem

+++ MentorBook.Web/Controllers/UserController.cs
@@ -93,10 +93,9 @@ namespace MentorBook.Web.Controllers
             dbUser.LastName = user.LastName;
             dbUser.Email = user.Email;
             dbUser.Phone = user.Phone;
-            dbUser.CurrentTownId = (int)user.CurrentTownId;
-            dbUser.HomeTownId = (int)user.HomeTownId;
+            dbUser.CurrentTownId = user.CurrentTownId;
+            dbUser.HomeTownId = user.HomeTownId;
             dbUser.DateOfBirth = user.DateOfBirth;
-            
 
             bool result = _userService.InsertUser(dbUser);
